refactor: adjust Models
